Split off some CWM widgets to separate files.
